+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 * Added `js_callscoped` to call c function in stack isolated environment, similar to js_call, but without intermediate object.
 * Added `js_seal`, `js_freeze`, `js_issealed`, `js_isfrozen` functions to change object extension abilities.
 * Added `js_catch_exit`, `js_exit`, `js_atexit` to issue and handle graceful script exit requests.
+* Added `js_loadstringE` to load code with predefined environment variables (push object, call func).